LayoutTests/imported/w3c/web-platform-tests/content-security-policy/reporting/report-cross-origin-no-cookies.sub-expected.txt

CONSOLE MESSAGE: Refused to load http://localhost:8800/content-security-policy/support/fail.png because it does not appear in the img-src directive of the Content Security Policy.
Blocked access to external URL http://www1.localhost:8800/content-security-policy/support/report.py?op=put&reportID=01584473-7c07-4eb2-a55c-55c074e5c3a0


PASS Image should not load
FAIL Violation report status OK. assert_equals: No such report. expected "" but got "false"
PASS Test report cookies.
